+++ CHANGELOG.md
@@ -1,8 +1,13 @@
 ## Unreleased
 
+
+### New Features
 - `from_regex/1`:
     - Add support for `:caseless` (`/i`) modifier
     - Do not generate full text matching strings unless `^` and/or `$` are included
+
+### Breaking changes
+- Remove `timezone/0` generator
 
 ## 0.1.0 [2026-03-25]
 Initial version